customizing the meals screen
