+++ MealzApp/app/src/main/java/com/mistpaag/mealzapp/ui/MainActivity.kt
@@ -1,0 +1,18 @@
+package com.mistpaag.mealzapp.ui
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import com.mistpaag.mealzapp.ui.meals.MealsCategoriesScreen
+import com.mistpaag.mealzapp.ui.theme.MealzAppTheme
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            MealzAppTheme {
+                MealsCategoriesScreen()
+            }
+        }
+    }
+}
