--- MealzApp/app/src/main/java/com/mistpaag/mealzapp/ui/meals/MainActivity.kt
@@ -1,48 +1,0 @@
-package com.mistpaag.mealzapp.ui.meals
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.viewModels
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.material.Text
-import androidx.compose.runtime.*
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.lifecycle.viewmodel.compose.viewModel
-import com.mistpaag.mealzapp.model.response.MealResponse
-import com.mistpaag.mealzapp.ui.theme.MealzAppTheme
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            MealzAppTheme {
-                MealsCategoriesScreen()
-            }
-        }
-    }
-}
-
-@Composable
-fun MealsCategoriesScreen() {
-    val viewModel : MealsCategoriesViewModel = viewModel()
-    val meals = viewModel.mealsState.value
-
-    LazyColumn {
-        items(meals) { meal ->
-            Text(text = meal.name)
-        }
-    }
-
-}
-
-@Preview(showBackground = true)
-@Composable
-fun DefaultPreview() {
-    MealzAppTheme {
-        MealsCategoriesScreen()
-    }
-}
